renamed DeclarationType module to TypeDeclaration
