+++ RAGESS/RAGESS/Sources/TypeDeclaration/Function.swift
