+++ RAGESS/RAGESS/Sources/TypeDeclaration/StructObject.swift
